+++ src/com/braintreegateway/BraintreeGateway.java
@@ -46,7 +46,7 @@ import com.braintreegateway.util.TrUtil;
  */
 public class BraintreeGateway {
 
-    public static final String VERSION = "2.2.1";
+    public static final String VERSION = "2.3.0";
 
     private Configuration configuration;
     private Environment environment;
@@ -123,6 +123,10 @@ public class BraintreeGateway {
         return new TransactionGateway(http, configuration);
     }
 
+    public TransparentRedirectGateway transparentRedirect() {
+        return new TransparentRedirectGateway(http, configuration);
+    }
+    
     /**
      * Returns encoded transparent redirect data for the given {@link Request} and redirect URL
      * @param trData the transparent redirect data as a {@link Request} object.