+++ src/com/braintreegateway/CustomerSearchRequest.java
@@ -1,7 +1,0 @@
-package com.braintreegateway;
-
-public class CustomerSearchRequest extends SearchRequest {
-    public MultipleValueNode<CustomerSearchRequest> ids() {
-        return new MultipleValueNode<CustomerSearchRequest>("ids", this);
-    }
-}